# chore: optimise schema generation function