--- a/cmd/benchmark/Dockerfile
+++ b/cmd/benchmark/Dockerfile
@@ -12,7 +12,7 @@ COPY . .
 
 RUN go build -o throttling ./cmd/benchmark/throttling/
 
-FROM alpine:3.21.0
+FROM alpine:3.21.3
 
 COPY --from=builder workdir/throttling .