# Description

Optimise the Schema Generation Function in batch router. Attached are the benchmarks for the same with before and after implementations

The main bottleneck in Drain function was with respect to config and we fixed that as part of another PR

Generate Schema Map Benchmark

```
BenchmarkGenerateSchemaMap
BenchmarkGenerateSchemaMap/Sequential_Small_1K_Jobs_5Tables_10Columns
BenchmarkGenerateSchemaMap/Sequential_Small_1K_Jobs_5Tables_10Columns-12         	     507	   2798334 ns/op	 2715276 B/op	   72027 allocs/op
BenchmarkGenerateSchemaMap/PerTable_Small_1K_Jobs_5Tables_10Columns
BenchmarkGenerateSchemaMap/PerTable_Small_1K_Jobs_5Tables_10Columns-12           	     828	   1412946 ns/op	 4050449 B/op	    8112 allocs/op
BenchmarkGenerateSchemaMap/Sequential_Medium_10K_Jobs_10Tables_20Columns
BenchmarkGenerateSchemaMap/Sequential_Medium_10K_Jobs_10Tables_20Columns-12      	      27	  43566591 ns/op	43756744 B/op	 1140084 allocs/op
BenchmarkGenerateSchemaMap/PerTable_Medium_10K_Jobs_10Tables_20Columns
BenchmarkGenerateSchemaMap/PerTable_Medium_10K_Jobs_10Tables_20Columns-12        	      52	  21844792 ns/op	74738436 B/op	  100304 allocs/op
BenchmarkGenerateSchemaMap/Sequential_Large_100K_Jobs_20Tables_30Columns
BenchmarkGenerateSchemaMap/Sequential_Large_100K_Jobs_20Tables_30Columns-12      	       2	 578630396 ns/op	718992364 B/op	15600215 allocs/op
BenchmarkGenerateSchemaMap/PerTable_Large_100K_Jobs_20Tables_30Columns
BenchmarkGenerateSchemaMap/PerTable_Large_100K_Jobs_20Tables_30Columns-12        	       4	 271754427 ns/op	1423165464 B/op	 1201113 allocs/op
BenchmarkGenerateSchemaMap/Sequential_Small_100000_Jobs_100Tables_300Columns
BenchmarkGenerateSchemaMap/Sequential_Small_100000_Jobs_100Tables_300Columns-12  	       1	5107076000 ns/op	5355176472 B/op	124201550 allocs/op
BenchmarkGenerateSchemaMap/PerTable_Small_100000_Jobs_100Tables_300Columns
BenchmarkGenerateSchemaMap/PerTable_Small_100000_Jobs_100Tables_300Columns-12    	       1	2399475750 ns/op	11841683648 B/op	 1803405 allocs/op

```

## Linear Ticket

Fixes PIPE-1963

## Security

- [ ] The code changed/added as part of this pull request won't create any security issues with how the software is being used.